refactor(devops): Remove openzeppelin and use waffle as compiler
