--- .openzeppelin/project.json
@@ -1,30 +1,0 @@
-{
-  "manifestVersion": "2.2",
-  "contracts": {
-    "DXswapFactory": "DXswapFactory",
-    "DXswapPair": "DXswapPair",
-    "DXswapERC20": "DXswapERC20"
-  },
-  "dependencies": {},
-  "name": "dxswap-core",
-  "version": "1.0.0",
-  "compiler": {
-    "compilerSettings": {
-      "evmVersion": "istanbul",
-      "optimizer": {
-        "enabled": true,
-        "runs": "200"
-      }
-    },
-    "manager": "openzeppelin",
-    "typechain": {
-      "enabled": true,
-      "outDir": "./build/types/",
-      "target": "web3-v1"
-    },
-    "solcVersion": "0.5.16",
-    "artifactsDir": "build/contracts",
-    "contractsDir": "contracts"
-  },
-  "telemetryOptIn": true
-}
